Fix GCC compiler warnings

+++ source/kernel/src/threads.unix.cpp
@@ -315,7 +315,7 @@ bool xke::initThreadSubsystem(){
 	xke::thread_data.threads[0] = pthread_self();
 	xke::THIS_THREAD_INDEX = 0;
 	// start at i = 1, thread 0 is the calling thread
-	for(uint i = 1; i < num_threads; ++i){
+	for(u64 i = 1; i < num_threads; ++i){
 
 		if(0 != pthread_create(&xke::thread_data.threads[i],
 		                       &attribs,
